requirements_utf8.txt
alembic==1.18.4
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bidict==0.23.1
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
granian==2.7.2
greenlet==3.3.2
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.18.1
numpy==2.4.3
oauth2client==4.1.3
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
Pyrebase4==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-jwt==4.1.0
python-multipart==0.0.22
python-socketio==5.16.1
pytz==2026.1.post1
redis==7.4.0
referencing==0.37.0
reflex==0.8.28.post1
reflex-hosting-cli==0.1.62
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==1.0.0
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
watchdog==6.0.0
watchfiles==1.1.1
wrapt==2.1.2
wsproto==1.3.2
